Changes -------------- 1. Separate CodeTable to a new file 2. Add functional extension of code tables 3. Add code table displayable strings property
